Explicit casting for coordinates.

+++ util/VoxelConverterTool/src/Pipeline/VoxToFaces.cpp
@@ -102,7 +102,13 @@ namespace VoxelConverterTool{
                 }
                 //Submit this face
 
-                const WrappedFaceContainer c = {x, y, z, 1, 1, 1, v, ambientMask, f, 0};
+                const WrappedFaceContainer c = {
+                    static_cast<uint8>(x),
+                    static_cast<uint8>(y),
+                    static_cast<uint8>(z),
+                    1, 1, 1, v, ambientMask,
+                    static_cast<uint8>(f), 0
+                };
                 //WrappedFace face = _wrapFace(c);
                 faces.outFaces.push_back(c);
             }
